refactor: move chords maps to utils
