+++ utils/chords.js
@@ -1,0 +1,68 @@
+export const transposeMap = [
+  ['Am', 'A#m', 'Bm', 'Cm', 'C#m', 'Dm', 'D#m', 'Em', 'Fm', 'F#m', 'Gm', 'G#m'],
+  ['C', 'C#', 'D', 'D#', 'E', 'F', 'F#', 'G', 'G#', 'A', 'A#', 'B'],
+  ['Cmaj7', 'C#maj7', 'Dmaj7', 'D#maj7', 'Emaj7', 'Fmaj7', 'F#maj7', 'Gmaj7', 'G#maj7', 'Amaj7', 'A#maj7', 'Bmaj7'],
+  ['C6', 'C#6', 'D6', 'D#6', 'E6', 'F6', 'F#6', 'G6', 'G#6', 'A6', 'A#6', 'B6'],
+  ['C+7', 'C#+7', 'D+7', 'D#+7', 'E+7', 'F+7', 'F#+7', 'G+7', 'G#+7', 'A+7', 'A#+7', 'B+7'],
+  ['Csus2', 'C#sus2', 'Dsus2', 'D#sus2', 'Esus2', 'Fsus2', 'F#sus2', 'Gsus2', 'G#sus2', 'Asus2', 'A#sus2', 'Bsus2'],
+  ['Csus4', 'C#sus4', 'Dsus4', 'D#sus4', 'Esus4', 'Fsus4', 'F#sus4', 'Gsus4', 'G#sus4', 'Asus4', 'A#sus4', 'Bsus4'],
+  ['Csus2/A#', 'C#sus2/B', 'Dsus2/C', 'D#sus2/C#', 'Esus2/D', 'Fsus2/D#', 'F#sus2/E', 'Gsus2/F', 'G#sus2/F#', 'Absus2/Gb', 'A#sus2/Ab', 'Bsus2/A'],
+  ['Cm/A#', 'C#m/B', 'Dm/C', 'D#m/C#', 'Em/D', 'Fm/D#', 'F#m/E', 'Gm/F', 'G#m/F#', 'Abm/Gb', 'A#m/G#', 'Bm/A'],
+  ['A5', 'A#5', 'B5', 'C5', 'C#5', 'D5', 'D#5', 'E5', 'F5', 'F#5', 'G5', 'G#5'],
+  ['A7', 'A#7', 'B7', 'C7', 'C#7', 'D7', 'D#7', 'E7', 'F7', 'F#7', 'G7', 'G#7'],
+  ['Am6', 'A#m6', 'Bm6', 'Cm6', 'C#m6', 'Dm6', 'D#m6', 'Em6', 'Fm6', 'F#m6', 'Gm6', 'G#m6'],
+  ['Am7', 'A#m7', 'Bm7', 'Cm7', 'C#m7', 'Dm7', 'D#m7', 'Em7', 'Fm7', 'F#m7', 'Gm7', 'G#m7'],
+  ['Am7-5', 'A#m7-5', 'Bm7-5', 'Cm7-5', 'C#m7-5', 'Dm7-5', 'D#m7-5', 'Em7-5', 'Fm7-5', 'F#m7-5', 'Gm7-5', 'G#m7-5']
+];
+
+export const chordNotesMap = {
+  'E': [40, 44, 47],
+  'E5': [40, 44, 47], // [40, 47],
+  'E7': [40, 44, 47], // [40, 44, 47, 50],
+  'Em': [40, 43, 47],
+  'Em7': [40, 43, 47], // [40, 43, 47, 50],
+  'C': [48, 52, 55], // [36, 40, 43]
+  'C5': [36, 40, 43], // [36, 43],
+  'C#7': [37, 41, 44], // [37, 41, 44, 47],
+  'Cm': [36, 39, 43],
+  'C#m': [37, 40, 44],
+  'Cmaj': [36, 40, 43], // [36, 40, 43, 47]
+  'Cmaj7': [36, 40, 43], // [36, 40, 43, 47]
+  'C#': [37, 41, 44],
+  'D': [38, 42, 45],
+  'D7': [38, 42, 45], // [38, 42, 45, 48]
+  'Dsus2': [38, 42, 45],
+  'D#': [39, 43, 46],
+  'D5': [38, 42, 45], // [38, 45]
+  'Dm': [50, 53, 57], // [38, 41, 45],
+  'Dm7': [50, 53, 57], // [50, 53, 57, 60]
+  'D#m': [39, 42, 46],
+  'Hm': [35, 38, 42],
+  'Bm': [35, 38, 42],
+  'H': [35, 39, 42], // [47, 51, 54]
+  'B': [35, 39, 42], // [47, 51, 54]
+  'H5': [35, 39, 42], // [35, 42]
+  'B5': [35, 39, 42], // [35, 42]
+  'B7': [35, 39, 42], // [47, 51, 54]
+  'H7': [35, 39, 42], // [47, 51, 54]
+  'G': [43, 47, 50],
+  'G5': [43, 47, 50], // [43, 50]
+  'Gm': [43, 46, 50],
+  'G6': [43, 47, 50], // [43, 47, 50, 52]
+  'G#': [44, 48, 51],
+  'G#7': [44, 48, 51], // [44, 48, 51, 54]
+  'F': [41, 45, 48],
+  'Fm': [41, 44, 48],
+  'F5': [41, 45, 48], // [41, 48]
+  'F#': [42, 46, 49],
+  'A': [45, 49, 52],
+  'A7': [45, 49, 52], // [45, 49, 52, 55]
+  'A5': [45, 49, 52], // [45, 52]
+  'Am': [45, 48, 52],
+  'Am7': [45, 48, 52], // [45, 48, 52, 55]
+  'A#': [46, 50, 53],
+  // дальше не точно
+  'F#m': [42, 45, 49],
+  'G#m': [44, 47, 51],
+  'A#m': [46, 49, 53],
+};
